Fixes Select Component initial value is always the first option
When using the SelectControl component, the initial value for the select
element upon page load uses the value of the first child option element,
rather than having no - or an empty - value. To deal with this, added
the first disabled option "- Select -". Since the 'SelectControl as
BaseComponent'  from '@wordpress/components' doesn't support the
'disabled' attribute for the <option> element, switched to using a
manual <select> element, instead of using the BaseComponent from the
library.

Also in the Components Demo wizard, updated more meaningful labels
and fixed the wrong use of onChange to handle event as argument,
rather than value.