perf: reduce JS bundle size
By using SplitChunksPlugin we can separate the reused code and reduce the bundle size by 80%

The downside is that now the public-path.js import has to be manually set in all wizards. This is because the array entry did not play well with chunk splitting. Webpack 5 has more options for entry definition, which might help, but:
1. we rely on calypso-build's version on webpack, which is now at 4.x
2. this manual import is not that big of a price to pay for an 80% deliverable code size reduction