feat(support): list user's support tickets (#534)
A big change here is that now a support ticket will be submitted on behalf of the authenticated WordPress.com user, not the WordPress instance user. If the emails of these users match then there's nothing to worry about. If not, you might not see some of the tickets in the tickets list. 

But we're working on improving the tickets list to show tickets related to the organization (a list of users), not a single user.