feat(connections): add MailChimp API key setting in connection screen
MailChimp API key is managed in the newsletter plugin while it's used in other plugins too. Move the
MailChimp API key option to the main plugin (this one).

BREAKING CHANGE: No breaking changes for this plugin, but for the plugins using the MailChimp API
Key option they need to use the new one from this commit instead.