fix: handle variable source IDs
Stripe Subscriptions created before the changes in
#2251 will not have a Source, but the customer will
have PaymentMethods. These changes ensure that these
customers' subscriptions can be migrated.

This also means that not all changes from #2251 are
necessary – Stripe Payment Gateway creates sources,
but will handle a PaymentMethod ID if set as the source
ID in the subscription metadata.