feat: remove new tab default on image credits (#2880)
* Removes new tab default on image credits

Image credits default to opening in a new tab. In general, opening in a new tab is actually not a great user experience, as it deprives the user of the freedom of whether to open it in the current tab or a new one.

There are also accessibility considerations, as opening in a new tab makes the site problematic for users with disabilities who use assistive technologies.

This handy guide explains why it’s not a great idea and details some narrow cases where it might be preferable:

https://css-tricks.com/use-target_blank/

The thinking on this has evolved fairly recently, but current best practices are to open most, if not all, links in the same tab (with possible exceptions for when it opens a document like a PDF or something else that takes you out of your regular browser experience). There's more about the approach in this article from DMXzone:

https://www.dmxzone.com/go/10742/free-beware-of-opening-links-in-a-new-window/

Publishers who insist on making this the default are welcome to install a separate plugin that enforces this behavior, as opposed to it being decided by the theme.