fix: use percentage rather than value for cover fee text (#3280)
This fixes an issue where we are using the cover fee display value in the cover fee name. Woo escapes HTML entities when outputting the cover fee name as a result, HTML tags are rendered on the frontend in the transaction order review table. We fix this by outputting just a string of the percentage rather than the display value.