fix(ras-acc): handle invalid removechild js error (#3281)
This resolves some js errors that appear whenever transitioning from the auth modal to the checkout modal. It looks like there is some logic that tries to remove a loading state spinner in the auth modal which throws an error since the modal is closed and transitioned to the checkout modal.