fix(ras-acc): redirect to homepage after logging out via my account (#3272)
This PR redirects readers to the homepage after logging out via my account. This is because we now provide a signin option from every other page and it's not necessary to visit the my-account page to login anymore when RAS is setup. This also handles cases where folks are logging in from the my account page, forcing a redirect to trigger the updated my account dashboard.